+++ TabbedTortoiseGit/ChangelogDialog.Designer.cs
@@ -37,9 +37,10 @@
             // 
             this.OkButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
             this.OkButton.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.OkButton.Location = new System.Drawing.Point(597, 476);
+            this.OkButton.Location = new System.Drawing.Point(696, 549);
+            this.OkButton.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.OkButton.Name = "OkButton";
-            this.OkButton.Size = new System.Drawing.Size(75, 23);
+            this.OkButton.Size = new System.Drawing.Size(88, 27);
             this.OkButton.TabIndex = 1;
             this.OkButton.Text = "OK";
             this.OkButton.UseVisualStyleBackColor = true;
@@ -48,9 +49,10 @@
             // 
             this.ShowChangelogOnUpdateCheck.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.ShowChangelogOnUpdateCheck.AutoSize = true;
-            this.ShowChangelogOnUpdateCheck.Location = new System.Drawing.Point(12, 480);
+            this.ShowChangelogOnUpdateCheck.Location = new System.Drawing.Point(14, 555);
+            this.ShowChangelogOnUpdateCheck.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.ShowChangelogOnUpdateCheck.Name = "ShowChangelogOnUpdateCheck";
-            this.ShowChangelogOnUpdateCheck.Size = new System.Drawing.Size(157, 17);
+            this.ShowChangelogOnUpdateCheck.Size = new System.Drawing.Size(171, 19);
             this.ShowChangelogOnUpdateCheck.TabIndex = 2;
             this.ShowChangelogOnUpdateCheck.Text = "Show changelog on update";
             this.ShowChangelogOnUpdateCheck.UseVisualStyleBackColor = true;
@@ -62,26 +64,27 @@
             | System.Windows.Forms.AnchorStyles.Right)));
             this.ChangelogText.BorderStyle = System.Windows.Forms.BorderStyle.None;
             this.ChangelogText.Cursor = System.Windows.Forms.Cursors.Default;
-            this.ChangelogText.Location = new System.Drawing.Point(12, 12);
+            this.ChangelogText.Location = new System.Drawing.Point(14, 14);
+            this.ChangelogText.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.ChangelogText.Name = "ChangelogText";
             this.ChangelogText.ReadOnly = true;
-            this.ChangelogText.Size = new System.Drawing.Size(660, 458);
+            this.ChangelogText.Size = new System.Drawing.Size(770, 528);
             this.ChangelogText.TabIndex = 3;
             this.ChangelogText.Text = "";
-            this.ChangelogText.WordWrap = false;
             // 
             // ChangelogDialog
             // 
             this.AcceptButton = this.OkButton;
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.CancelButton = this.OkButton;
-            this.ClientSize = new System.Drawing.Size(684, 511);
+            this.ClientSize = new System.Drawing.Size(798, 590);
             this.Controls.Add(this.ChangelogText);
             this.Controls.Add(this.ShowChangelogOnUpdateCheck);
             this.Controls.Add(this.OkButton);
+            this.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.MinimizeBox = false;
-            this.MinimumSize = new System.Drawing.Size(275, 175);
+            this.MinimumSize = new System.Drawing.Size(318, 196);
             this.Name = "ChangelogDialog";
             this.ShowInTaskbar = false;
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;