+++ TabbedTortoiseGit/Properties/Settings.Designer.cs
@@ -569,5 +569,17 @@ namespace TabbedTortoiseGit.Properties {
                 this["TreatUninitializedSubmodulesAsModified"] = value;
             }
         }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("")]
+        public string LastVersion {
+            get {
+                return ((string)(this["LastVersion"]));
+            }
+            set {
+                this["LastVersion"] = value;
+            }
+        }
     }
 }