+++ TabbedTortoiseGit/Resources/Changelog.md
@@ -2,6 +2,11 @@
 If you've updated Tabbed TortoiseGit but your settings were lost, go to Settings -> Restore Settings.
 This should restore your settings from the most recent settings file.
 
+# Version 0.37.0
+* Improved installation process.
+* Add current branch macro to custom actions.
+* Improve update checking.
+
 # Version 0.36.0
 * Handle uninitialized submodules as modified submodules for Fast(er) Submodule Update.
 * Check for TortoiseGitProc.exe in default location before checking PATH.